Jazzy Water
Jazzed up the water ~
I also tried to get the rock texture to load in correctly, but had to stop before I could get it to work.
(I feel like i was on the right path? But i ran out of time and commented out what i was working on, leaving the OG code active)